backend/requirements.txt
fastapi
uvicorn
typing_extensions
langchain
langchain-community
openai
python-dotenv
requests
langchain-deepseek
langchain-core
langchain-openai